Fix PR #4805 review issues in XLSX, PPTX, and CLI policy code
