+++ mineru/backend/utils/office_chart.py
@@ -1,3 +1,4 @@
+# Copyright (c) Opendatalab. All rights reserved.
 import re
 from io import BytesIO
