+++ mineru/cli/router.py
@@ -40,6 +40,13 @@ from mineru.cli.api_client import (
 )
 from mineru.cli.api_protocol import API_PROTOCOL_VERSION
 from mineru.cli.common import normalize_upload_filename
+from mineru.cli.public_http_client_policy import (
+    PUBLIC_HTTP_CLIENT_DISABLED_DETAIL,
+    configure_public_http_client_policy,
+    is_public_bind_host,
+    validate_public_http_client_request,
+    warn_if_public_http_client_policy as _warn_if_public_http_client_policy,
+)
 from mineru.cli.vlm_preload import build_local_api_cli_args
 from mineru.version import __version__
 
@@ -66,11 +73,6 @@ WORKER_REFRESH_INTERVAL_SECONDS = 2.0
 MIN_HEALTHY_PROCESSING_WINDOW_SIZE = 1
 MINERU_ROUTER_PUBLIC_BIND_EXPOSED_ENV = "MINERU_ROUTER_PUBLIC_BIND_EXPOSED"
 MINERU_ROUTER_ALLOW_PUBLIC_HTTP_CLIENT_ENV = "MINERU_ROUTER_ALLOW_PUBLIC_HTTP_CLIENT"
-PUBLIC_HTTP_CLIENT_DISABLED_DETAIL = (
-    "Publicly exposed API disables *-http-client backends and server_url by "
-    "default. Rebind to 127.0.0.1 or start with "
-    "--allow-public-http-client if you understand the SSRF risk."
-)
 
 
 def utc_now_iso() -> str:
@@ -127,51 +129,11 @@ def is_task_terminal(status: str) -> bool:
     return status in TASK_TERMINAL_STATES
 
 
-def is_public_bind_host(host: str) -> bool:
-    return host in {"0.0.0.0", "::"}
-
-
-def configure_public_http_client_policy(
-    app: FastAPI,
-    *,
-    public_bind_exposed: bool,
-    allow_public_http_client: bool,
-) -> None:
-    app.state.public_bind_exposed = public_bind_exposed
-    app.state.allow_public_http_client = allow_public_http_client
-
-
-def validate_public_http_client_request(
-    *,
-    public_bind_exposed: bool,
-    allow_public_http_client: bool,
-    backend: str,
-    server_url: str | None,
-) -> None:
-    if not public_bind_exposed or allow_public_http_client:
-        return
-    if backend.endswith("-http-client") or bool(server_url and server_url.strip()):
-        raise HTTPException(status_code=400, detail=PUBLIC_HTTP_CLIENT_DISABLED_DETAIL)
-
-
 def warn_if_public_http_client_policy(host: str, allow_public_http_client: bool) -> None:
-    if not is_public_bind_host(host):
-        return
-    if allow_public_http_client:
-        logger.warning(
-            "MinerU router is listening on {} with --allow-public-http-client enabled. "
-            "Requests may supply remote HTTP inference endpoints and turn the service "
-            "into an externally driven outbound request primitive, creating SSRF and "
-            "internal network probing risk.",
-            host,
-        )
-        return
-    logger.warning(
-        "MinerU router is listening on {}. Disabling *-http-client backends and "
-        "server_url by default because these inputs let callers choose remote HTTP "
-        "inference endpoints; when the API is publicly reachable, that creates SSRF "
-        "and internal network probing risk.",
-        host,
+    _warn_if_public_http_client_policy(
+        service_name="router",
+        host=host,
+        allow_public_http_client=allow_public_http_client,
     )
